[chatranks] Restore vanilla chat when disabled
This has been a request to revert back to vanilla when disabled for those who need it for various different Bots or to be able to use different packs that offer other experiences for chat ranks. This should help minimize conflicts but is still not guaranteed as a perfect solution since any other pack utilizing the `before ChatSend Event` can still cause issues if not handled properly.