Fix a lot of bugs with Channels
When you join a channel and leave you were not properly removed from playerChannelMap. This has been fixed.

When you join a channel and the owner deletes it you and the other members are not properly removed from playerChannelMap. This has been fixed.

Made cosmetic changes when channels are deleted to let members know it has been disbanded. Also included a message to let players know who invited them to their channel. Furthermore, I updated the join subcommand to let the player know that they have join the channel.