[phase] Track soul sand and account for slight phasing
Soul Sand has an effect that causes players to slightly sink into it. This is how the block causes players to move slower. There is a -0.125 difference in the Y coordinate when this happens. We track if its soul sand and we confirm that they have not exceeded this difference. Should they exceed that difference then they will be flagged.