[illegalitemsB] Temporarily ignore hoppers
There is a false flag triggered when using hoppers that can have Paradox believe it contains nested items at the time of it being placed in the world, however its not accurate solely do to the event handler that we are listening to. Since its an after event listener it gets the data about the blocks container after it is already placed and we need to know the information before. There will be future changes to the API that should allow this problem to be properly resolved. Until then ignore hoppers.