[showRules] Fix playersAwaitingResponse
This was originally implemented in 33c96ac4592c940c16645dc6569a72c41e06a7bd and it was intended to verify if the player was pending to prevent duplicating the form on the same player. This did it's job but later it was consequently broken in bcbb8a35d5c3558990b5df0910906225b2d44a7e and went unnoticed until now after a report was made regarding a problem with duplicated forms. This patch corrects the problem by ensuring players are managed accordingly to eliminate that duplication. They only need to be shown the form once, and we wait for a response code. From there we decide what to do after a response code is recieved.