[Registry] Append string to enum from MinecraftEntityTypes
Unknowing to me, it seems teh developers overlooked the registerEntityTypeDynamicProperties function when handling strings for the second arg. In most cases the developers have designed their API to accept strings with or without the prefix "minecraft:" but assumes it if not included. In this particular case, it is expected and there is no exclusion to the rule here. MinecraftENtityTypes is an enum in vanilla-data and the enum returns the string of that entity without the prefix. So instead of getting "minecraft:player" you would get "player". In most cases, this would be acceptable but not here so we append "minecraft:" at the beginning of the string to make the API happy.