Add 'minecraft/server' as a dependency in package.json
There seems to be a small bug with the overrides attribute where package-lock.json is only partially updated. Adding the package as a dependency, even though it is redundant, resolves the issue by ensuring package-lock.json is fully filled. The overrides attribute is still utilized to ensure that specific version is not overwritten by the dependency called on for 'minecraft/server-ui'.