[imprison] Increase height for imprisonment by one block
Although the height was sufficient for a player, there is a weird bug in the game where you begin to suffocate as though you are inside a block after logging off then returning. Seems to be an oversight, but increasing the height by one block prevents the bug from triggering.