Remove container for dependencies
I added this a while back because there was an issue with environments
being out of scope which would throw errors and this was the only way to
resolve the problem.

Returning back to this, I have found it is no longer a problem and I can
safely eliminate this container.